총 모음/backend/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
slowapi==0.1.9

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 캐싱
redis==5.0.1
hiredis==2.2.3

# ML/DL 프레임워크
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.3.1
huggingface-hub==0.20.2

# 한국어 NLP
konlpy==0.6.0

# 데이터 처리
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# 웹 스크래핑 (데이터 수집용)
beautifulsoup4==4.12.2
requests==2.31.0

# 유틸리티
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# 로깅 및 모니터링
python-json-logger==2.0.7

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 시각화 (평가용)
matplotlib==3.8.2
seaborn==0.13.0

# Gemini API
google-generativeai==0.3.2

# Google Cloud Platform
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0

# 개발 도구
black==23.11.0
flake8==6.1.0
mypy==1.7.1
